// server/src/generated/database.ts
export interface OddsHistoryRowRead {
"uid": string 
"event_key": string 
"sport_key": string 
"sport_nice": string 
"commence_time": number 
"home_team": string 
"away_team": string 
"teams": any 
"sites": any 
"sites_count": number 
"created_at": any 
"load_uid": string 
"market": string 
}
// ...
export interface LinesHistoryRowWrite {
"uid"?: string 
"event_key": string 
"sport_key": string 
"home_team": string 
"away_team": string 
"event_start_time": any 
"h2h_home"?: any | null
"h2h_away"?: any | null
"h2h_draw"?: any | null
"spread_home"?: any | null
"spread_away"?: any | null
"spread_home_vig"?: any | null
"spread_away_vig"?: any | null
"total"?: any | null
"total_over_vig"?: any | null
"total_under_vig"?: any | null
"load_uid": string 
"created_at"?: any 
}
// ...
export interface SportsRowWrite {
"key": string 
"active": boolean 
"group": string 
"details": string 
"title": string 
"has_outrights": boolean 
"created_at"?: any 
}